minor refactor of how we identify a message format. fixed #106
created `smsparser.getSMSFormat`, `isMuvukuFormat` and
`isTextformsFormat` so when a new message is added it gets a `format`
property now, which gets added to the data_record doc. so, added this
property to tests.   also consolidated getCallbackBody a little. Found a
bug where _.each function closure removed variables from scope we were
expecting to be there.