fixed tests for smsforms -> jsonforms refactor.
still some ugly parsing going on and it would be nice to streamline
tests, remove redundant tests and add specific coverage for all the form
field types and features.  will add these as separate issues. NOTE also
modified parser to not throw errors instead just return the raw value by
default and log a message if a field type is not handled. this is mainly
to avoid 500 errors because of a likely minor parsing issue.