rewrote getValues and added some unit tests. fixes spreadsheet output in certain cases.
for nyaya we had to include health_centers on related_entities on the
the data record since that is where reporting was coming from. this
causes some issues when clinic is present but null, breaking the
spreadsheet output.  fixed that so utils.getValues should be more
flexible with keys and better tested now.