reduced repetitive display on outgoing bulk messaging #333
Added temporary helper property to doc when it is passed to renderer
with `outgoing_messages` property looks something like this so messages
with the same `message`, `state`, `from` and `sent_by` values are
grouped together. Timestamp value remains on the recipient because that
depends on when the message was sent, they can be sent independently
especially if used with kujua-transport as opposed to SMSSync.

```
outgoing_messages = [
    {
        recipients: [
            {
                to: '+123',
                facility: <facility>,
                timestamp: <timestamp>,
                uuid: <uuid>,
            },
        ...
        ],
        sent_by: 'admin',
        from: '+998',
        state: 'sent',
        message: 'good morning'
    },
    ...
]
```

Removed messages param from `contact` template helper, it doesn't belong
there trying to do too much.

If there are 5 or less recipients we try to display them in the minimize
row otherwise we say "<length> recipients".

TODO expand recipient buttons to display additional information if
helpful, like phone number and timestamp/date when state was set.